Attempt to fix "User code threw uncaught exception" error
Before, the servos were initialized using "HardwareMap.get" which
returns a "HardwareDevice" object, then attempting to convert it to a
"Servo" object. That might be why the code was crashing. Anyway, it is
probably better to have an initialization function for servos, since we
have one for the motors. I made the changes to the
"DiamondBladeAutonomous" program too, since if that is the problem it
would affect that program too.