Preliminary shooting functions, comments, cleanup
changed the code to have one function pull back the shooter, lock it,
and unwind the shooter and another function to release the shooter.
shooter cannot be released until it is done unwinding. THIS HAS NOT BEEN
TESTED. added a bunch of comments. and moved a few lines of code (put
all the motor initializations together, for example)